+++ tests/ConfirmEqualTest.cs
@@ -1,3 +1,4 @@
+using System;
 using Confirma.Attributes;
 using Confirma.Exceptions;
 using Confirma.Extensions;
@@ -30,13 +31,17 @@ public static class ConfirmEqualTest
 	[TestCase(2f, 2)]
 	public static void ConfirmEqual_WhenNotEqual(object o1, object o2)
 	{
-		ConfirmExtensions.ConfirmThrows<ConfirmAssertException>(() => o1.ConfirmEqual(o2));
+		Action action = () => o1.ConfirmEqual(o2);
+
+		action.ConfirmThrows<ConfirmAssertException>();
 	}
 
 	[TestCase(new string[] { "Hello,", "world!" }, new string[] { "Hi,", "world!" })]
 	public static void ConfirmEqual_WhenArrNotEqual(object[] o1, object[] o2)
 	{
-		ConfirmExtensions.ConfirmThrows<ConfirmAssertException>(() => o1.ConfirmEqual(o2));
+		Action action = () => o1.ConfirmEqual(o2);
+
+		action.ConfirmThrows<ConfirmAssertException>();
 	}
 
 	[TestCase(1, 2)]
@@ -61,12 +66,16 @@ public static class ConfirmEqualTest
 	[TestCase(2f, 2f)]
 	public static void ConfirmNotEqual_WhenEqual(object o1, object o2)
 	{
-		ConfirmExtensions.ConfirmThrows<ConfirmAssertException>(() => o1.ConfirmNotEqual(o2));
+		Action action = () => o1.ConfirmNotEqual(o2);
+
+		action.ConfirmThrows<ConfirmAssertException>();
 	}
 
 	[TestCase(new string[] { "Hello,", "world!" }, new string[] { "Hello,", "world!" })]
 	public static void ConfirmNotEqual_WhenArrEqual(object[] o1, object[] o2)
 	{
-		ConfirmExtensions.ConfirmThrows<ConfirmAssertException>(() => o1.ConfirmNotEqual(o2));
+		Action action = () => o1.ConfirmNotEqual(o2);
+
+		action.ConfirmThrows<ConfirmAssertException>();
 	}
 }